Move LibraryCommander.{joinLibrary, leaveLibrary, updateLibraryMembershipAccess} to LibraryMembershipCommander
