+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/AuthCommander.scala
@@ -116,7 +116,7 @@ class AuthCommander @Inject() (
     s3ImageStore: S3ImageStore,
     postOffice: LocalPostOffice,
     inviteCommander: InviteCommander,
-    libraryCommander: LibraryCommander,
+    libraryMembershipCommander: LibraryMembershipCommander,
     orgMembershipCommander: OrganizationMembershipCommander,
     orgInviteCommander: OrganizationInviteCommander,
     implicit val publicIdConfig: PublicIdConfiguration,
@@ -459,7 +459,7 @@ class AuthCommander @Inject() (
     // Abstracting away errors and manually reporting. If someone needs the specific error, feel free to change the signature.
     Library.decodePublicId(libPubId).map { libId =>
       implicit val context = HeimdalContext(Map())
-      libraryCommander.joinLibrary(userId, libId, authToken).fold(
+      libraryMembershipCommander.joinLibrary(userId, libId, authToken).fold(
         { libFail =>
           airbrake.notify(s"[finishSignup] lib-auto-join failed. $libFail")
           false
