+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/LibraryCommander.scala
@@ -1,42 +1,24 @@
 package com.keepit.commanders
 
-import com.google.inject.{ ImplementedBy, Inject, Provider }
-import com.keepit.abook.ABookServiceClient
-import com.keepit.commanders.emails.{ EmailOptOutCommander, LibraryInviteEmailSender }
+import com.google.inject.{ ImplementedBy, Inject }
 import com.keepit.common.akka.SafeFuture
-import com.keepit.common.cache._
-import com.keepit.common.concurrent.FutureHelpers
+
 import com.keepit.common.core._
 import com.keepit.common.crypto.PublicIdConfiguration
 import com.keepit.common.db.Id
-import com.keepit.common.db.slick.DBSession.{ RSession, RWSession }
+import com.keepit.common.db.slick.DBSession.RWSession
 import com.keepit.common.db.slick.Database
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.logging.Logging
-import com.keepit.common.mail.{ BasicContact, EmailAddress }
-import com.keepit.common.performance.{ AlertingTimer, StatsdTiming }
-import com.keepit.common.social.BasicUserRepo
-import com.keepit.common.store.{ ImageSize, S3ImageStore }
 import com.keepit.common.time._
-import com.keepit.eliza.{ ElizaServiceClient, LibraryPushNotificationCategory, PushNotificationExperiment, UserPushNotificationCategory }
-import com.keepit.heimdal.{ HeimdalContext, HeimdalContextBuilderFactory, HeimdalServiceClient }
-import com.keepit.model
-import com.keepit.model.LibraryAccess.READ_ONLY
+import com.keepit.heimdal.HeimdalContext
 import com.keepit.model.LibrarySpace.{ OrganizationSpace, UserSpace }
 import com.keepit.model._
-import com.keepit.notify.model.Recipient
-import com.keepit.notify.model.event.{ LibraryNewKeep, OwnedLibraryNewCollaborator, OwnedLibraryNewFollower }
 import com.keepit.search.SearchServiceClient
-import com.keepit.social.{ BasicNonUser, BasicUser }
-import com.keepit.typeahead.KifiUserTypeahead
 import com.kifi.macros.json
-import org.joda.time.DateTime
 import play.api.http.Status._
-import play.api.libs.json._
 
-import scala.collection.parallel.ParSeq
 import scala.concurrent._
-import scala.concurrent.duration._
 
 @json case class MarketingSuggestedLibrarySystemValue(
   id: Id[Library],
@@ -58,8 +40,6 @@ trait LibraryCommander {
   def deleteLibrary(libraryId: Id[Library], userId: Id[User])(implicit context: HeimdalContext): Option[LibraryFail]
   def canMoveTo(userId: Id[User], libId: Id[Library], to: LibrarySpace): Boolean
   def createReadItLaterLibrary(userId: Id[User]): Library
-  def joinLibrary(userId: Id[User], libraryId: Id[Library], authToken: Option[String] = None, subscribed: Option[Boolean] = None)(implicit eventContext: HeimdalContext): Either[LibraryFail, (Library, LibraryMembership)]
-  def leaveLibrary(libraryId: Id[Library], userId: Id[User])(implicit eventContext: HeimdalContext): Either[LibraryFail, Unit]
   def copyKeepsFromCollectionToLibrary(userId: Id[User], libraryId: Id[Library], tagName: Hashtag)(implicit context: HeimdalContext): Either[LibraryFail, (Seq[Keep], Seq[(Keep, LibraryError)])]
   def moveKeepsFromCollectionToLibrary(userId: Id[User], libraryId: Id[Library], tagName: Hashtag)(implicit context: HeimdalContext): Either[LibraryFail, (Seq[Keep], Seq[(Keep, LibraryError)])]
   def copyKeeps(userId: Id[User], toLibraryId: Id[Library], keeps: Set[Keep], withSource: Option[KeepSource])(implicit context: HeimdalContext): (Seq[Keep], Seq[(Keep, LibraryError)])
@@ -68,7 +48,6 @@ trait LibraryCommander {
   def trackLibraryView(viewerId: Option[Id[User]], library: Library)(implicit context: HeimdalContext): Unit
   def updateLastEmailSent(userId: Id[User], keeps: Seq[Keep]): Unit
   def updateSubscribedToLibrary(userId: Id[User], libraryId: Id[Library], subscribedToUpdatesNew: Boolean): Either[LibraryFail, LibraryMembership]
-  def updateLibraryMembershipAccess(requestUserId: Id[User], libraryId: Id[Library], targetUserId: Id[User], newAccess: Option[LibraryAccess]): Either[LibraryFail, LibraryMembership]
   def unsafeTransferLibrary(libraryId: Id[Library], newOwner: Id[User]): Library
 }
 
@@ -78,28 +57,20 @@ class LibraryCommanderImpl @Inject() (
     libraryMembershipRepo: LibraryMembershipRepo,
     libraryAliasRepo: LibraryAliasRepo,
     libraryInviteRepo: LibraryInviteRepo,
-    libraryInviteCommander: LibraryInviteCommander,
     librarySubscriptionCommander: LibrarySubscriptionCommander,
     organizationMembershipCommander: OrganizationMembershipCommander,
     userRepo: UserRepo,
-    basicUserRepo: BasicUserRepo,
     keepRepo: KeepRepo,
     keepCommander: KeepCommander,
     keepToCollectionRepo: KeepToCollectionRepo,
     ktlRepo: KeepToLibraryRepo,
     ktlCommander: KeepToLibraryCommander,
-    libraryAccessCommander: LibraryAccessCommander,
     countByLibraryCache: CountByLibraryCache,
     typeaheadCommander: TypeaheadCommander,
-    kifiUserTypeahead: KifiUserTypeahead,
     collectionRepo: CollectionRepo,
-    s3ImageStore: S3ImageStore,
     airbrake: AirbrakeNotifier,
     searchClient: SearchServiceClient,
-    elizaClient: ElizaServiceClient,
     libraryAnalytics: LibraryAnalytics,
-    libraryImageCommander: LibraryImageCommander,
-    kifiInstallationCommander: KifiInstallationCommander,
     implicit val defaultContext: ExecutionContext,
     implicit val publicIdConfig: PublicIdConfiguration,
     clock: Clock) extends LibraryCommander with Logging {
@@ -523,182 +494,6 @@ class LibraryCommanderImpl @Inject() (
     readItLaterLib
   }
 
-  private def notifyOwnerOfNewFollowerOrCollaborator(newFollowerId: Id[User], lib: Library, access: LibraryAccess): Unit = SafeFuture {
-    val (follower, owner, lotsOfFollowers) = db.readOnlyReplica { implicit session =>
-      val follower = userRepo.get(newFollowerId)
-      val owner = basicUserRepo.load(lib.ownerId)
-      val lotsOfFollowers = libraryMembershipRepo.countMembersForLibrarySince(lib.id.get, DateTime.now().minusDays(1)) > 2
-      (follower, owner, lotsOfFollowers)
-    }
-    val (title, category, message) = if (access == LibraryAccess.READ_WRITE) {
-      // This should be changed to library_collaborated but right now iOS skips categories it doesn't know.
-      ("New Library Collaborator", NotificationCategory.User.LIBRARY_FOLLOWED, s"${follower.firstName} ${follower.lastName} is now collaborating on your Library ${lib.name}")
-    } else {
-      ("New Library Follower", NotificationCategory.User.LIBRARY_FOLLOWED, s"${follower.firstName} ${follower.lastName} is now following your Library ${lib.name}")
-    }
-    val libImageOpt = libraryImageCommander.getBestImageForLibrary(lib.id.get, ProcessedImageSize.Medium.idealSize)
-    elizaClient.sendGlobalNotification( //push sent
-      userIds = Set(lib.ownerId),
-      title = title,
-      body = message,
-      linkText = s"See ${follower.firstName}’s profile",
-      linkUrl = s"https://www.kifi.com/${follower.username.value}",
-      imageUrl = s3ImageStore.avatarUrlByUser(follower),
-      sticky = false,
-      category = category,
-      unread = !lotsOfFollowers, // if not a lot of recent followers, notification is marked unread
-      extra = Some(Json.obj(
-        "follower" -> BasicUser.fromUser(follower),
-        "library" -> Json.toJson(LibraryNotificationInfoBuilder.fromLibraryAndOwner(lib, libImageOpt, owner))
-      ))
-    ) map { _ =>
-        if (!lotsOfFollowers) {
-          val canSendPush = kifiInstallationCommander.isMobileVersionEqualOrGreaterThen(lib.ownerId, KifiAndroidVersion("2.2.4"), KifiIPhoneVersion("2.1.0"))
-          if (canSendPush) {
-            val pushCat = category match {
-              case NotificationCategory.User.LIBRARY_COLLABORATED => UserPushNotificationCategory.NewLibraryCollaborator
-              case _ => UserPushNotificationCategory.NewLibraryFollower
-            }
-            elizaClient.sendUserPushNotification(
-              userId = lib.ownerId,
-              message = message,
-              recipient = follower,
-              pushNotificationExperiment = PushNotificationExperiment.Experiment1,
-              category = pushCat)
-          }
-        }
-      }
-    if (access == LibraryAccess.READ_WRITE) {
-      elizaClient.sendNotificationEvent(OwnedLibraryNewCollaborator(
-        Recipient(lib.ownerId),
-        currentDateTime,
-        newFollowerId,
-        lib.id.get
-      ))
-    } else {
-      elizaClient.sendNotificationEvent(OwnedLibraryNewFollower(
-        Recipient(lib.ownerId),
-        currentDateTime,
-        newFollowerId,
-        lib.id.get
-      ))
-    }
-  }
-
-  def joinLibrary(userId: Id[User], libraryId: Id[Library], authToken: Option[String] = None, subscribed: Option[Boolean] = None)(implicit eventContext: HeimdalContext): Either[LibraryFail, (Library, LibraryMembership)] = {
-    val (lib, inviteList, existingActiveMembership) = db.readOnlyMaster { implicit s =>
-      val lib = libraryRepo.get(libraryId)
-      val tokenInvites = if (authToken.isDefined) {
-        libraryAccessCommander.getValidLibInvitesFromAuthToken(libraryId, authToken)
-      } else Seq.empty
-      val libInvites = libraryInviteRepo.getWithLibraryIdAndUserId(libraryId, userId)
-      val allInvites = tokenInvites ++ libInvites
-      val existingActiveMembership = libraryMembershipRepo.getWithLibraryIdAndUserId(libraryId, userId)
-      (lib, allInvites, existingActiveMembership)
-    }
-
-    val isSystemGeneratedLibrary = lib.kind == LibraryKind.SYSTEM_MAIN || lib.kind == LibraryKind.SYSTEM_SECRET
-    val userCanJoinLibraryWithoutInvite = libraryAccessCommander.canViewLibrary(Some(userId), lib, authToken) // uses a db session
-
-    if (isSystemGeneratedLibrary) {
-      Left(LibraryFail(FORBIDDEN, "cant_join_system_generated_library"))
-    } else if (!userCanJoinLibraryWithoutInvite && inviteList.isEmpty && existingActiveMembership.isEmpty) {
-      // private library & no library invites with matching authtoken
-      Left(LibraryFail(FORBIDDEN, "cant_join_library_without_an_invite"))
-    } else {
-      val maxAccess: LibraryAccess = {
-        val orgMemberAccess: Option[LibraryAccess] = if (lib.isSecret) None else lib.organizationId.flatMap { orgId =>
-          lib.organizationMemberAccess.filter { _ => organizationMembershipCommander.getMembership(orgId, userId).isDefined }
-        }
-        (inviteList.map(_.access).toSet ++ orgMemberAccess).maxOpt.getOrElse(LibraryAccess.READ_ONLY)
-      }
-      val (updatedLib, updatedMem) = db.readWrite(attempts = 3) { implicit s =>
-        val updatedMem = libraryMembershipRepo.getWithLibraryIdAndUserId(libraryId, userId, None) match {
-          case None =>
-            val subscribedToUpdates = subscribed.getOrElse(maxAccess == LibraryAccess.READ_WRITE)
-            log.info(s"[joinLibrary] New membership for $userId. New access: $maxAccess. $inviteList")
-            val mem = libraryMembershipRepo.save(LibraryMembership(libraryId = libraryId, userId = userId, access = maxAccess, lastJoinedAt = Some(clock.now), subscribedToUpdates = subscribedToUpdates))
-            notifyOwnerOfNewFollowerOrCollaborator(userId, lib, maxAccess) // todo, bad, this is in a db transaction and side effects
-            mem
-          case Some(mem) =>
-            val maxWithExisting = if (mem.state == LibraryMembershipStates.ACTIVE) Seq(maxAccess, mem.access).max else maxAccess
-            val subscribedToUpdates = subscribed.getOrElse(maxWithExisting == LibraryAccess.READ_WRITE || mem.subscribedToUpdates)
-            log.info(s"[joinLibrary] Modifying membership for ${mem.userId} / $userId. Old access: ${mem.access} (${mem.state}), new: $maxWithExisting. $maxAccess, $inviteList")
-            libraryMembershipRepo.save(mem.copy(access = maxWithExisting, state = LibraryMembershipStates.ACTIVE, lastJoinedAt = Some(clock.now), subscribedToUpdates = subscribedToUpdates))
-        }
-
-        inviteList.foreach { inv =>
-          // Only update invitations to a specific user. If it's to a specific recipient. Otherwise, leave it open for others.
-          if (inv.userId.isDefined || inv.emailAddress.isDefined) {
-            libraryInviteRepo.save(inv.copy(state = LibraryInviteStates.ACCEPTED))
-          }
-        }
-
-        val invitesToAlert = inviteList.filterNot(_.inviterId == lib.ownerId)
-        if (invitesToAlert.nonEmpty) {
-          val invitee = userRepo.get(userId)
-          val owner = basicUserRepo.load(lib.ownerId)
-          libraryInviteCommander.notifyInviterOnLibraryInvitationAcceptance(invitesToAlert, invitee, lib, owner) // todo, bad, this is in a db transaction and side effects
-        }
-
-        val updatedLib = libraryRepo.save(lib.copy(memberCount = libraryMembershipRepo.countWithLibraryId(libraryId)))
-        (updatedLib, updatedMem)
-      }
-      updateLibraryJoin(userId, lib, updatedMem, eventContext)
-      Right((updatedLib, updatedMem))
-    }
-  }
-
-  private def updateLibraryJoin(userId: Id[User], library: Library, membership: LibraryMembership, eventContext: HeimdalContext): Future[Unit] = SafeFuture {
-    val libraryId = library.id.get
-    libraryAnalytics.acceptLibraryInvite(userId, library, eventContext)
-    libraryAnalytics.followLibrary(userId, library, eventContext)
-    searchClient.updateLibraryIndex()
-    if (LibraryAccess.collaborativePermissions.contains(membership.access)) {
-      refreshLibraryCollaboratorsTypeahead(libraryId)
-    }
-  }
-
-  def leaveLibrary(libraryId: Id[Library], userId: Id[User])(implicit eventContext: HeimdalContext): Either[LibraryFail, Unit] = {
-    db.readOnlyMaster { implicit s =>
-      libraryMembershipRepo.getWithLibraryIdAndUserId(libraryId, userId, None)
-    } match {
-      case None => Right((): Unit)
-      case Some(mem) if mem.access == LibraryAccess.OWNER => Left(LibraryFail(BAD_REQUEST, "cannot_leave_own_library"))
-      case Some(mem) =>
-        val lib = db.readWrite { implicit s =>
-          libraryMembershipRepo.save(mem.copy(state = LibraryMembershipStates.INACTIVE))
-          val lib = libraryRepo.get(libraryId)
-          libraryRepo.save(lib.copy(memberCount = libraryMembershipRepo.countWithLibraryId(libraryId)))
-          lib
-        }
-        SafeFuture {
-          convertKeepOwnershipToLibraryOwner(userId, lib)
-          libraryAnalytics.unfollowLibrary(userId, lib, eventContext)
-          searchClient.updateLibraryIndex()
-          if (LibraryAccess.collaborativePermissions.contains(mem.access)) {
-            refreshLibraryCollaboratorsTypeahead(libraryId)
-          }
-        }
-        Right((): Unit)
-    }
-  }
-
-  // TODO(ryan): is this actually necessary anymore? Check with Léo to see if Search needs it
-  // We may not need to have this concept of "keep ownership" anymore. You can be the author of
-  // a keep, which is its own thing. You can have access to edit a keep, which is also its own
-  // thing.
-  private def convertKeepOwnershipToLibraryOwner(userId: Id[User], library: Library) = {
-    db.readWrite { implicit s =>
-      keepRepo.getByUserIdAndLibraryId(userId, library.id.get).foreach { keep =>
-        keepCommander.changeOwner(keep, library.ownerId)
-      }
-      ktlRepo.getByUserIdAndLibraryId(userId, library.id.get).foreach {
-        ktlCommander.changeOwner(_, library.ownerId)
-      }
-    }
-  }
-
   def copyKeepsFromCollectionToLibrary(userId: Id[User], libraryId: Id[Library], tagName: Hashtag)(implicit context: HeimdalContext): Either[LibraryFail, (Seq[Keep], Seq[(Keep, LibraryError)])] = {
     db.readOnlyMaster { implicit s =>
       collectionRepo.getByUserAndName(userId, tagName)
@@ -846,72 +641,5 @@ class LibraryCommanderImpl @Inject() (
         Right(updatedMembership)
       }
     }
-  }
-
-  ///////////////////
-  // Collaborators!
-  ///////////////////
-
-  def updateLibraryMembershipAccess(requestUserId: Id[User], libraryId: Id[Library], targetUserId: Id[User], newAccess: Option[LibraryAccess]): Either[LibraryFail, LibraryMembership] = {
-    if (newAccess.isDefined && newAccess.get == LibraryAccess.OWNER) {
-      Left(LibraryFail(BAD_REQUEST, "cannot_change_access_to_owner"))
-    } else {
-      db.readOnlyMaster { implicit s =>
-        val membershipMap = libraryMembershipRepo.getWithLibraryIdAndUserIds(libraryId, Set(requestUserId, targetUserId))
-        val library = libraryRepo.get(libraryId)
-        (membershipMap.get(requestUserId), membershipMap.get(targetUserId), library)
-      } match {
-        case (None, _, _) =>
-          Left(LibraryFail(NOT_FOUND, "request_membership_not_found"))
-        case (_, None, _) =>
-          Left(LibraryFail(NOT_FOUND, "target_membership_not_found"))
-        case (Some(mem), Some(targetMem), _) if targetMem.access == LibraryAccess.OWNER =>
-          Left(LibraryFail(BAD_REQUEST, "cannot_change_owner_access"))
-
-        case (Some(requesterMem), Some(targetMem), library) =>
-
-          if ((requesterMem.isOwner && !targetMem.isOwner) || // owners can edit anyone except themselves
-            (requesterMem.isCollaborator && !targetMem.isOwner) || // a collaborator can edit anyone (but the owner). Collaborator cannot invite others to collaborate if the library does not allow collaborators to invite
-            (requesterMem.isFollower && requesterMem.userId == targetMem.userId)) { // a follower can only edit herself
-            db.readWrite { implicit s =>
-              newAccess match {
-                case None =>
-                  SafeFuture { convertKeepOwnershipToLibraryOwner(targetMem.userId, library) }
-                  Right(libraryMembershipRepo.save(targetMem.copy(state = LibraryMembershipStates.INACTIVE)))
-                case Some(newAccess) if requesterMem.isCollaborator && newAccess == LibraryAccess.READ_WRITE && library.whoCanInvite == Some(LibraryInvitePermissions.OWNER) =>
-                  log.warn(s"[updateLibraryMembership] invalid permission ${requesterMem} trying to change membership ${targetMem} to ${newAccess} when library has invite policy ${library.whoCanInvite}")
-                  Left(LibraryFail(FORBIDDEN, "invalid_collaborator_permission"))
-                case Some(newAccess) =>
-                  val newSubscription = if (newAccess == LibraryAccess.READ_WRITE) true else targetMem.subscribedToUpdates // auto subscribe to updates if a collaborator
-                  val inviter = userRepo.get(requestUserId)
-                  val libOwner = basicUserRepo.load(library.ownerId)
-                  val updatedTargetMembership = libraryMembershipRepo.save(targetMem.copy(access = newAccess, subscribedToUpdates = newSubscription, state = LibraryMembershipStates.ACTIVE))
-
-                  if (inviter.id.get.id != targetUserId.id) {
-                    libraryInviteCommander.notifyInviteeAboutInvitationToJoinLibrary(inviter, library, libOwner, Map(targetUserId -> updatedTargetMembership))
-                  }
-
-                  Right(updatedTargetMembership)
-              }
-            } tap {
-              // Unless we're just kicking out a follower, the set of collaborators has changed.
-              case Right(updatedMembership) if !(updatedMembership.isFollower && updatedMembership.state == LibraryMembershipStates.INACTIVE) => {
-                SafeFuture { refreshLibraryCollaboratorsTypeahead(library.id.get) }
-              }
-              case _ => //
-            }
-          } else { // invalid permissions
-            log.warn(s"[updateLibraryMembership] invalid permission ${requesterMem} trying to change membership ${targetMem} to ${newAccess}")
-            Left(LibraryFail(FORBIDDEN, "invalid_permissions"))
-          }
-      }
-    }
-  }
-
-  private def refreshLibraryCollaboratorsTypeahead(libraryId: Id[Library]): Future[Unit] = {
-    val collaboratorIds = db.readOnlyMaster { implicit session =>
-      libraryMembershipRepo.getCollaboratorsByLibrary(Set(libraryId)).get(libraryId).toSet.flatten
-    }
-    kifiUserTypeahead.refreshByIds(collaboratorIds.toSeq)
   }
 }
